--- web/src/routes/experiments/+page.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-  import PublicExperiments from "$lib/components/public-experiments.svelte";
-  let { data } = $props();
-</script>
-
-<PublicExperiments experiments={data.experiments} />
-
